refactor(config): consolidate settings file creation primitives in permissions module
